backend/requirements.txt
google-genai
fastapi
pydantic
uvicorn